Patches to make Gwydion Dylan build on RedHat 5 systems, using modern 2.0.x kernels. From now on, you'll have to run autoconf before building.

+++ mindy/compat/std-bstring.h
@@ -1,4 +1,4 @@
-extern void	bcopy _ANSI_ARGS_((const void *, void *, int));
-extern int	bcmp _ANSI_ARGS_((const void *, const void *, int));
-extern void	bzero _ANSI_ARGS_((void *, int));
+extern void	bcopy _ANSI_ARGS_((const void *, void *, size_t));
+extern int	bcmp _ANSI_ARGS_((const void *, const void *, size_t));
+extern void	bzero _ANSI_ARGS_((void *, size_t));
 extern void	blkclr _ANSI_ARGS_((void *, int));
